requirements.txt
Flask==2.3.3
requests==2.31.0
Pillow==10.1.0
gunicorn==21.2.0
Werkzeug==2.3.7